move valid_token_side_effect to utils, direct dict access,
